requirements.txt
requests>=2.30.0
ibm-watson>=7.2.0
ibm-cloud-sdk-core>=3.16.0
newspaper3k>=0.2.8
flask>=2.3.0
